+++ Assets/OxGKit/SingletonSystem/Scripts/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## CHANGELOG
 
+## [1.0.1] - 2025-04-08
+- Added InitInstance method.
+- Modified the default value of dontDestroyOnLoad in MonoSingleton to true.
+
 ## [1.0.0] - 2025-03-19
 - Added SingletonSystem.